Sync #62 from devel to testing
Fix the last qos in a namespace cannot be deleted

If the last qos in a namespace is deleted, server will not send any
information about that to the agent. Thus the agent should keep which
namespaces it is hosting to be able to delete qdiscs which should be
removed.

Fixes: redmine #8954

Signed-off-by: Hunt Xu <mhuntxu@gmail.com>
(cherry picked from commit 86b28cbf0d2246261729fe16a896eb43b9512a22)
Signed-off-by: Hunt Xu <mhuntxu@gmail.com>