metering: always setup metering iptables rules for routers
The deployment of the metering iptables rules depends on routers
connecting to external networks with router ports. As we have introduced
EayunStack floatingip mechanism, this dependency should be removed. And
we can now do metering for a floating IP even though its associated
router is not connected to the external network with a router port.

Fixes: redmine #9992

Signed-off-by: Hunt Xu <mhuntxu@gmail.com>